Move setPropsFromUrl to AAQStep component
