+++ kitsune/questions/static/questions/js/components/TopicSelector.jsx
@@ -4,18 +4,10 @@ import AAQStep from './AAQStep.jsx';
 import AAQActions from '../actions/AAQActions.es6.js';
 import aaqGa from '../utils/aaqGa.es6.js';
 import UrlActions from '../../../sumo/js/actions/UrlActions.es6.js';
-import UrlStore from '../../../sumo/js/stores/UrlStore.es6.js';
 
 const topics = JSON.parse(document.querySelector('.data[name=topics]').innerHTML);
 
 export default class TopicSelector extends AAQStep {
-
-  setPropsFromUrl() {
-    let urlData = UrlStore.get('pathProps');
-    if (urlData.product) {
-      AAQActions.setProduct(urlData.product);
-    }
-  }
 
   componentWillMount() {
     if (!this.props.question.product) {
